perf: remove extra cast

+++ src/OracleMiddleware/oracles/ChainlinkDataStreamsOracle.sol
@@ -119,11 +119,13 @@ abstract contract ChainlinkDataStreamsOracle is IOracleMiddlewareErrors, IChainl
             revert OracleMiddlewareWrongBidPrice(verifiedReport.bid);
         }
 
+        // The following values (price, ask, bid) have been validated to be greater than 0,
+        // making the casting to uint192 safe.
         return FormattedDataStreamsPrice({
-            timestamp: uint256(verifiedReport.observationsTimestamp),
-            price: uint256(uint192(verifiedReport.price)),
-            ask: uint256(uint192(verifiedReport.ask)),
-            bid: uint256(uint192(verifiedReport.bid))
+            timestamp: verifiedReport.observationsTimestamp,
+            price: uint192(verifiedReport.price),
+            ask: uint192(verifiedReport.ask),
+            bid: uint192(verifiedReport.bid)
         });
     }
